+++ src/index.ts
@@ -151,6 +151,7 @@ export { hasClearContrast } from "./inspectors/hasClearContrast";
 export { isDark } from "./inspectors/isDark";
 export { isDull } from "./inspectors/isDull";
 export { isLight } from "./inspectors/isLight";
+export { isMidtone } from "./inspectors/isMidtone";
 export { isMonochrome } from "./inspectors/isMonochome";
 export { lightModeContrastRatio } from "./inspectors/lightModeContrastRatio";
 export { luma } from "./inspectors/luma";